Add window selection
The process is slowing down a bit, and most of it is just trying to
figure out what will and will not work, but hey, that's how it goes.

I think that I'm getting pretty close to having something a little more
interesting to test. I'd like to try out end-to-end tests, but that's
probably not terribly possible. We'll see.